Main: Cleanup walkdir test flow
Cross check not just if it is a dir, but also check if its a normal
file or not. Needed bcas walkdir doesnt walk symlinks and symlink
is not identified as a dir or a regular file.

Rearrange to first get the type of path and then log it also along
with path.