Ast: Update Idents DB/Map from GenDecl(Value/TypeSpec), FuncDecl
Setup to update identifiers/symbols map from GenDecl and FuncDecl
instead of directly from lower Ident object. This is because, one
can get associated Comments at the GenDecl/FuncDecl but not at
Ident object level.

NOTE1: This commit uses ast.IsExported, just to show it and also
bcas I wrongly forgot that Name in TypeSpec is a Ident, will be
changing it in next commit.

NOTE2: The way ident is updated if already present in the map, may
not actually work, bcas the Ident value stored in the map is a
struct, which gets copied into a new struct var ident when assigned.
Will fix it later, showing just for making this point and a associated
test next.