Main+: Store MatchMode also as a uint and use it
As MatchMode check is done at the core, in the innerloop, of the
goapropos logic, so instead of doing a string based comparison
each time, do a faster integer based comparsion.

At same time allow enduser to specify the matchmode as a friendly
string, which will be converted to the uint based MatchMode type.