Ast: IdentyStats type with helper functions
Converted the identifiers related stats struct into a Type and added
the String method from Stringer interface to allow %v of fmt printing
the contents of the struct with tags for the fields.

Also added a delta summary helper to give the difference between
the generic/all identifiers and sum of specific types of identifiers
like type, val(const/var), func.

Inturn looking at the huge diff shown by the prints and inturn
looking at prints related to individual identifiers and the src,
realised the in a way obvious, if one thinks about it, that generic
identifiers node callback from AST will capture all identifiers
whether being defined by the go file being inspected, or being used
by it.

TODO: This also means that, given that the go/ast | go/parser |
go/token etal checks for whether a symbol is exported or not based
on whether it starts with a capital letter or not, this catch all
identifiers path was adding exported symbols of entities used
from other packages in a go source file, as also belonging to
the package to which the go source file being inspected belonged.

Need to fix this issue. Which should be fine, given that ValueSpec,
TypeSpec and FuncDecl paths trigger only for functions defined
within the go source file being inspected. So the redundant catch
all identifiers path needs to be jetisoned out, as it feeds own
and others identifiers into the apropos's symbols database.