Ast: Link block level comments wrt Value/Type decls to them
Now any comment given in the go source file at a block level of
a bunch of consts or types or vars is associated with the members
of that block. So that a search for a part of that comment will
show all the members of that block.

Misc: Madam Secretary series giving nice company in the background
As also the new Star Trek with Cmdr Pike.