DBPlus: GoApropos wrt S_IRUSR returns results
Using GoApropos to search for S_IRUSR gave hits wrt syscall, unix
and windows. Also grep seems to indicate syscall, inturn seem to
have some skeleton wrt different unixes and windows wrt this. Am
being lazy, havent actually looked into the code in the source,
beyond the grep.

So use it. And while at it, also tack on write permission also,
bcas we want to be able to update/rewrite the db cache, after it
has been created, say when a new version of the golang language
suite is installed or so.