DB: Explicit MarshalJSON method for my Ident type - P1
A skeleton towards implementing a MarshalJSON logic for my Ident
type, because, otherwise the default json.Marshal is generating a
empty "{}" for Ident values.

As some characters like rune of \n or " or so need to be handled
before storing a string in JSON, so initially thought of doing
this string marshalling myself and even put logic for handling
\n rune as the string with runes for \ and n. But then " needs
to be handled and then maybe something else, dont remember now,
have to refresh on json spec once.

Rather have decided to take the easy way out for now by using
the json packages Marshal on string value, to do the above transforms,
in the next commit.