DB:+: Replace Ident struct with identCmt string
Need to see if this makes the saving and loading of symbols db
faster or not. Potentially it should make it a little bit faster,
if not anything else, as It also avoids the simple yet clumsy
conversion path I did wrt Ident struct.

Update: usecache path has seems to finish the job 25% faster wrt
time.