MultDB: Update db_find for Multiple DBs world/flow
This wont be used perse with the different individual dbs in the
multiple db world, but only with the final integrated top level
db. But still make it work with the db passed rather than assuming
a global db, which gives the flexibility to work with individual
dbs if required in future.