Go Struct Caps Named fields and Json Export
Renamed the fields in DBEntry to start with Capital letter, so that
go will allow those fields to be accessed by code from other modules
like json in this case.

Rather this is what was causing the DBEntrys' being a NO SHOW in
the cache files.

With the renaming there is no longer a need for custom Marshal and
Unmarshal JSON methods wrt DBEntry.