Cache: Use ModTimeStamp of Src directory as Cache Version
Chances are runtime.Version got from within the go based program
executable will be fixed, as they are statically linked by default.

So this cant be used to identify if a new version of the go lang
suite and or go source package has been installed or not.

So now it uses the modifcation timestamp wrt the go source directory
being used, as the version associated with the cache. So if the
source directory gets updated/modified, this timestamp should
change and thus inturn, the cache should get autoupdated in autocache
mode.

However note that if one changes contents of some subdirs within
the go source dir, then the src dir's modtime maynot change.