DB: Cleanup prints, ReUse TheDB for matching packages
Make the dbprints more flexible and add a version for symbols.

Move filtering out of packages to a seperate dbfilter helper.

Instead of using a seperate MatchingPkgs, now use new instances of
TheDB itself to store matching package and symbols info in db_find.
The idea is to allow more info wrt matching symbols and packages to
be captured and processed as required later.