Compare: Oops had forgotten about prepare
In default mode of program, the matchmode is contains and in this
case the match pattern string was passing through match_prepare,
but the check string was not passing through match_prepare, so
the Matchok was failing, bcas of case based mismatch.

With the check string passing through match_prepare, to help with
above, if the match pattern doesnt go through the same wrt matchmode
of regexp, then it will fail, due ot case based mismatch.

This change should help resolve this.