Compare+: MatcherCreate either succeeds or panics
The RE version of matcher is created using MustCompile, so there
is no need to pass around a failure, if any, wrt re to the callers.

This simplifies the flow in general, where ever matcher is used.

At same time, now if a invalid re pattern is passed to matcher,
it will panic the program, which is fine for this programs usecase.

Also db_find was previously just dropping the error if any with
matcher creation, which would have inturn lead to panic, when the
nil matcher would have been used further down in db_find logic.
This change will make it panic, where the actual issue occurs ie
during creation.