Compare: Extending a pointer type using methods not done
One can extend a type using methods and inturn pass the type being
extended as a pointer to that type, as part of the methods special
argument (ie reciever type).

However one cant extend a type which is defined to represent a
pointer to some other type, using methods.

NOTE: THis is just to capture same here, will not work