Compare: Interfaces-Methods-Types and Pointers
THis allows regexp.Regexp to be type aliased and inturn that type
alias to implement methods of a interface, but by defining the
methods to work on a pointer of the new type alias.

Inturn a *regexp.Regexp is converted / casted into value storing
/ representing the interface in question.

As these new methods work on pointer of the type alias and thus
inturn the pointer of regexp.Regexp, so no costly copying is
done wrt the regexp.Regexp struct when these methods are called
from a inner loop of a logic which loops a lot.