Ast:PkgBasePath make it os agnostic
Dont hardcode path seperator.

Convert the path seps if any wrt pkgBasePath+pkgName to '/',
irrespective of the os. So that what is maintained in the
cache and or shown to user always uses '/' similar to how
one uses the equivalent when importing.

Rename to pkgname_with_basepath.

Update the version