DB:SymbolsUpdate: Try avoid duplicating info in type string
If the same symbol name is defined in a package multiple times
as different types (maybe a symbol name which is on its own as well
as as part of some interface or so or ???, need to look into this later).

Inturn if the same type is seen again, then try avoid duplicating
the type tag within the type string in DBSymbolInfo struct for that
symbol.

NOTE: As different go routines may be working on different files
of a package at the same time, so the same symbol wrt a package
may be part of different maps associated with the individual
go routines, and inturn while merging the maps the type assocated
with symbols will not be single type tag, but could be multiple,
this is not handled currently.